requirements.txt
alembic==1.16.4
asyncpg>=0.27.0
email_validator==2.2.0
fastapi==0.116.1
httpx==0.28.1
Jinja2==3.1.6
Mako==1.3.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
SQLAlchemy==2.0.43
starlette==0.47.2
uvicorn==0.35.0
